qt gui: use double instead of float in automatic data type translation
When translating a SC slot into a QVariant, always translate a Float
into a double, since there is no way to differentiate between a float
and a double in SC.

Accordingly, all the invokable methods of widgets have to change
signature to use double instead of float, or the automatic signature
matching will fail.

+++ QtCollider/widgets/QcNumberBox.cpp
@@ -170,14 +170,14 @@ void QcNumberBox::setMaxDecimals( int d )
     setValue( _value ); // round current value
 }
 
-void QcNumberBox::increment( float factor )
+void QcNumberBox::increment( double factor )
 {
   if( !isReadOnly() || _valueType != Number ) return;
   setValue( value() + (step * factor) );
   Q_EMIT( action() );
 }
 
-void QcNumberBox::decrement( float factor )
+void QcNumberBox::decrement( double factor )
 {
   if( !isReadOnly() || _valueType != Number ) return;
   setValue( value() - (step * factor) );
