+++ QtCollider/widgets/QcRangeSlider.h
@@ -60,8 +60,8 @@ class QcRangeSlider :
     QSize minimumSizeHint() const;
 
   public Q_SLOTS:
-    void increment( float factor );
-    void decrement( float factor );
+    void increment( double factor );
+    void decrement( double factor );
 
   Q_SIGNALS:
     void action();
